port ls to new list stream formatting
