+++ crates/nu-command/src/viewers/table.rs
@@ -1,22 +1,17 @@
-use lscolors::Style;
 use nu_color_config::{get_color_config, style_primitive};
-use nu_engine::{column::get_columns, env_to_string, CallExt};
+use nu_engine::{column::get_columns, CallExt};
 use nu_protocol::{
     ast::{Call, PathMember},
     engine::{Command, EngineState, Stack, StateWorkingSet},
-    format_error, Category, Config, DataSource, Example, IntoPipelineData, ListStream,
-    PipelineData, PipelineMetadata, RawStream, ShellError, Signature, Span, SyntaxShape, Value,
+    format_error, Category, Config, Example, IntoPipelineData, ListStream, PipelineData, RawStream,
+    ShellError, Signature, Span, SyntaxShape, Value,
 };
 use nu_table::{Alignments, StyledString, TableTheme, TextStyle};
-use nu_utils::get_ls_colors;
+use terminal_size::{Height, Width};
+
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Instant;
-use std::{
-    path::PathBuf,
-    sync::atomic::{AtomicBool, Ordering},
-};
-use terminal_size::{Height, Width};
-use url::Url;
 
 const STREAM_PAGE_SIZE: usize = 1000;
 const STREAM_TIMEOUT_CHECK_INTERVAL: usize = 100;
@@ -136,24 +131,17 @@ impl Command for Table {
                     metadata: None,
                 })
             }
-            PipelineData::Value(Value::List { vals, .. }, metadata) => handle_row_stream(
+            PipelineData::Value(Value::List { vals, .. }, _metadata) => handle_row_stream(
                 engine_state,
                 stack,
                 ListStream::from_stream(vals.into_iter(), ctrlc.clone()),
                 call,
                 row_offset,
                 ctrlc,
-                metadata,
             ),
-            PipelineData::ListStream(stream, metadata) => handle_row_stream(
-                engine_state,
-                stack,
-                stream,
-                call,
-                row_offset,
-                ctrlc,
-                metadata,
-            ),
+            PipelineData::ListStream(stream, _metadata) => {
+                handle_row_stream(engine_state, stack, stream, call, row_offset, ctrlc)
+            }
             PipelineData::Value(Value::Record { cols, vals, .. }, ..) => {
                 let mut output = vec![];
 
@@ -195,14 +183,13 @@ impl Command for Table {
                 let base_pipeline = val.to_base_value(span)?.into_pipeline_data();
                 self.run(engine_state, stack, call, base_pipeline)
             }
-            PipelineData::Value(Value::Range { val, .. }, metadata) => handle_row_stream(
+            PipelineData::Value(Value::Range { val, .. }, _metadata) => handle_row_stream(
                 engine_state,
                 stack,
                 ListStream::from_stream(val.into_range_iter(ctrlc.clone())?, ctrlc.clone()),
                 call,
                 row_offset,
                 ctrlc,
-                metadata,
             ),
             x => Ok(x),
         }
@@ -247,102 +234,7 @@ fn handle_row_stream(
     call: &Call,
     row_offset: usize,
     ctrlc: Option<Arc<AtomicBool>>,
-    metadata: Option<PipelineMetadata>,
 ) -> Result<nu_protocol::PipelineData, nu_protocol::ShellError> {
-    let stream = match metadata {
-        Some(PipelineMetadata {
-            data_source: DataSource::Ls,
-        }) => {
-            let config = engine_state.config.clone();
-            let ctrlc = ctrlc.clone();
-            let ls_colors_env_str = match stack.get_env_var(engine_state, "LS_COLORS") {
-                Some(v) => Some(env_to_string("LS_COLORS", &v, engine_state, stack)?),
-                None => None,
-            };
-            let ls_colors = get_ls_colors(ls_colors_env_str);
-            let show_clickable_links = config.show_clickable_links_in_ls;
-
-            ListStream::from_stream(
-                stream.map(move |(mut x, _)| match &mut x {
-                    Value::Record { cols, vals, .. } => {
-                        let mut idx = 0;
-
-                        while idx < cols.len() {
-                            if cols[idx] == "name" {
-                                if let Some(Value::String { val: path, span }) = vals.get(idx) {
-                                    match std::fs::symlink_metadata(&path) {
-                                        Ok(metadata) => {
-                                            let style = ls_colors.style_for_path_with_metadata(
-                                                path.clone(),
-                                                Some(&metadata),
-                                            );
-                                            let ansi_style = style
-                                                .map(Style::to_crossterm_style)
-                                                // .map(ToNuAnsiStyle::to_nu_ansi_style)
-                                                .unwrap_or_default();
-                                            let use_ls_colors = config.use_ls_colors;
-
-                                            let full_path = PathBuf::from(path.clone())
-                                                .canonicalize()
-                                                .unwrap_or_else(|_| PathBuf::from(path));
-                                            let full_path_link = make_clickable_link(
-                                                full_path.display().to_string(),
-                                                Some(&path.clone()),
-                                                show_clickable_links,
-                                            );
-
-                                            if use_ls_colors {
-                                                vals[idx] = Value::String {
-                                                    val: ansi_style
-                                                        .apply(full_path_link)
-                                                        .to_string(),
-                                                    span: *span,
-                                                };
-                                            }
-                                        }
-                                        Err(_) => {
-                                            let style = ls_colors.style_for_path(path.clone());
-                                            let ansi_style = style
-                                                .map(Style::to_crossterm_style)
-                                                // .map(ToNuAnsiStyle::to_nu_ansi_style)
-                                                .unwrap_or_default();
-                                            let use_ls_colors = config.use_ls_colors;
-
-                                            let full_path = PathBuf::from(path.clone())
-                                                .canonicalize()
-                                                .unwrap_or_else(|_| PathBuf::from(path));
-                                            let full_path_link = make_clickable_link(
-                                                full_path.display().to_string(),
-                                                Some(&path.clone()),
-                                                show_clickable_links,
-                                            );
-
-                                            if use_ls_colors {
-                                                vals[idx] = Value::String {
-                                                    val: ansi_style
-                                                        .apply(full_path_link)
-                                                        .to_string(),
-                                                    span: *span,
-                                                };
-                                            }
-                                        }
-                                    }
-                                }
-                            }
-
-                            idx += 1;
-                        }
-
-                        x
-                    }
-                    _ => x,
-                }),
-                ctrlc,
-            )
-        }
-        _ => stream,
-    };
-
     let head = call.head;
     let width_param: Option<i64> = call.get_flag(engine_state, stack, "width")?;
 
@@ -364,30 +256,6 @@ fn handle_row_stream(
         span: head,
         metadata: None,
     })
-}
-
-fn make_clickable_link(
-    full_path: String,
-    link_name: Option<&str>,
-    show_clickable_links: bool,
-) -> String {
-    // uri's based on this https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
-
-    if show_clickable_links {
-        format!(
-            "\x1b]8;;{}\x1b\\{}\x1b]8;;\x1b\\",
-            match Url::from_file_path(full_path.clone()) {
-                Ok(url) => url.to_string(),
-                Err(_) => full_path.clone(),
-            },
-            link_name.unwrap_or(full_path.as_str())
-        )
-    } else {
-        match link_name {
-            Some(link_name) => link_name.to_string(),
-            None => full_path,
-        }
-    }
 }
 
 fn convert_to_table(
